requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
PyJWT==1.4.2
python-dotenv==0.18.0
six==1.16.0
SQLAlchemy==1.4.21
uuid==1.30
Werkzeug==2.0.1
